requirements.txt
langchain==0.3.14
langchain-groq==0.2.4
langchain-core==0.3.33
python-telegram-bot==21.9
python-dotenv==1.0.1
web3==7.6.0
apscheduler==3.10.4
sqlalchemy==2.0.36
aiosqlite==0.20.0
